# backend/openapi.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/reports": {
      "post": {
        "operationId": "ReportsController_create",
        "parameters": [],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateReportDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "รับเรื่องแล้ว",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReportResponseDto"
                    }
                  }
                }
              }
            }
          },
          "404": {
            "description": "ไม่พบสิ่งที่รายงาน",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "รายงานเรื่องนี้ไว้แล้ว",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "รายงานเนื้อหาหรือผู้ใช้ที่ไม่เหมาะสม",
        "tags": [
          "reports"
        ]
      },
      "get": {
        "description": "เฉพาะบุคลากรและผู้ดูแลระดับองค์กร",
        "operationId": "ReportsController_list",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "OPEN",
              "type": "string",
              "enum": [
                "OPEN",
                "RESOLVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "target_kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "REEL",
                "POST",
                "MESSAGE",
                "COMMENT",
                "USER"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReportResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "ไม่มีสิทธิ์ดูคิวร้องเรียน",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "คิวเรื่องร้องเรียนสำหรับผู้ดูแล (เก่าสุดก่อน)",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/mine": {
      "get": {
        "operationId": "ReportsController_listMine",
        "parameters": [
          {
            "name": "page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "default": 1,
              "example": 2,
              "type": "number"
            }
          },
          {
            "name": "per_page",
            "required": false,
            "in": "query",
            "schema": {
              "minimum": 1,
              "maximum": 100,
              "default": 20,
              "example": 50,
              "type": "number"
            }
          },
          {
            "name": "status",
            "required": false,
            "in": "query",
            "schema": {
              "default": "OPEN",
              "type": "string",
              "enum": [
                "OPEN",
                "RESOLVED",
                "REJECTED"
              ]
            }
          },
          {
            "name": "target_kind",
            "required": false,
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "REEL",
                "POST",
                "MESSAGE",
                "COMMENT",
                "USER"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "สำเร็จ พร้อมข้อมูลแบ่งหน้า",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data",
                    "meta"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "type": "array",
                      "items": {
                        "$ref": "#/components/schemas/ReportResponseDto"
                      }
                    },
                    "meta": {
                      "$ref": "#/components/schemas/PaginationMeta"
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "เรื่องที่ฉันเคยรายงาน และสถานะล่าสุด",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/{id}": {
      "patch": {
        "operationId": "ReportsController_resolve",
        "parameters": [
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ResolveReportDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "สำเร็จ",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "data"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": true
                    },
                    "data": {
                      "$ref": "#/components/schemas/ReportResponseDto"
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "ไม่มีสิทธิ์ปิดเรื่อง",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "409": {
            "description": "เรื่องนี้ถูกปิดไปแล้ว",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "success",
                    "error"
                  ],
                  "properties": {
                    "success": {
                      "type": "boolean",
                      "example": false
                    },
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message"
                      ],
                      "properties": {
                        "code": {
                          "type": "string",
                          "example": "not_found"
                        },
                        "message": {
                          "type": "string",
                          "example": "ไม่พบคลิปนี้ อาจถูกลบไปแล้ว"
                        },
                        "details": {
                          "type": "array",
                          "items": {
                            "type": "string"
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "summary": "ปิดเรื่อง — จัดการแล้ว หรือไม่เข้าข่าย",
        "tags": [
          "reports"
        ]
      }
    },
...
  },
  "info": {
    "title": "CS Nexus — aie4-social-reels",
    "description": "ระบบย่อยภายใต้ CSMJU2030 · ตัวตนมาจาก API Gateway ผ่าน header X-User-Id / X-Layer1-Role / X-Faculty",
    "version": "1.0.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "PaginationMeta": {
        "type": "object",
        "properties": {
          "current_page": {
            "type": "number",
            "example": 2
          },
          "per_page": {
            "type": "number",
            "example": 50
          },
          "total_pages": {
            "type": "number",
            "example": 50
          },
          "total_items": {
            "type": "number",
            "example": 2480
          }
        },
        "required": [
          "current_page",
          "per_page",
          "total_pages",
          "total_items"
        ]
      },
...
      "ReportResponseDto": {
        "type": "object",
        "properties": {
          "id": {
            "type": "string"
          },
          "target_kind": {
            "type": "string",
            "enum": [
              "REEL",
              "POST",
              "MESSAGE",
              "COMMENT",
              "USER"
            ]
          },
          "target_id": {
            "type": "string"
          },
          "reason": {
            "type": "string"
          },
          "status": {
            "type": "string",
            "enum": [
              "OPEN",
              "RESOLVED",
              "REJECTED"
            ]
          },
          "reporter_username": {
            "type": "string",
            "description": "ผู้รายงาน — ผู้ดูแลเห็นได้ เพราะการรายงานเท็จซ้ำ ๆ ต้องตามตัวได้"
          },
          "resolved_by_username": {
            "type": "object",
            "nullable": true
          },
          "resolved_at": {
            "type": "object",
            "nullable": true
          },
          "created_at": {
            "type": "string"
          }
        },
        "required": [
          "id",
          "target_kind",
          "target_id",
          "reason",
          "status",
          "reporter_username",
          "resolved_by_username",
          "resolved_at",
          "created_at"
        ]
      },
      "CreateReportDto": {
        "type": "object",
        "properties": {
          "target_kind": {
            "type": "string",
            "enum": [
              "REEL",
              "POST",
              "MESSAGE",
              "COMMENT",
              "USER"
            ],
            "example": "POST"
          },
          "target_id": {
            "type": "string",
            "example": "9f1c2b3a-0000-4000-8000-000000000000",
            "description": "id ของสิ่งที่รายงาน · ถ้าเป็น USER ให้ใส่ username"
          },
          "reason": {
            "type": "string",
            "example": "โพสต์นี้มีข้อความคุกคามรุ่นน้อง"
          }
        },
        "required": [
          "target_kind",
          "target_id",
          "reason"
        ]
      },
      "ResolveReportDto": {
        "type": "object",
        "properties": {
          "status": {
            "type": "string",
            "enum": [
              "RESOLVED",
              "REJECTED"
            ],
            "description": "RESOLVED = จัดการแล้ว · REJECTED = พิจารณาแล้วไม่เข้าข่าย"
          },
          "note": {
            "type": "string",
            "description": "บันทึกของผู้ดูแล เก็บลง audit log"
          }
        },
        "required": [
          "status"
        ]
      },
...
    }
  }
}